requirements.txt
spotipy
requests
PyYAML
deemix
python-dotenv
